moved notifications to entrantmain
