+++ Trojan0Project/app/src/main/java/com/example/trojan0project/EntrantMain.java
@@ -2,7 +2,13 @@ package com.example.trojan0project;
 
 import static android.content.Intent.getIntent;
 
+import android.Manifest;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -10,6 +16,8 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 import com.bumptech.glide.Glide;
 import com.google.firebase.firestore.DocumentSnapshot;
@@ -57,6 +65,30 @@ public class EntrantMain extends AppCompatActivity {
 
         loadProfileData();
 
+        // Fetch and display notifications for the device
+        Notification notificationHelper = new Notification();
+        notificationHelper.getNotificationsForDevice(this, deviceId);
+        // Request POST_NOTIFICATIONS permission
+        requestNotificationPermission();
+        // Call createNotificationChannel to ensure the channel is created on compatible devices
+        createNotificationChannel(this);
+
+        /*
+        String deviceId = "8fb329762643a6cb";
+
+        // Sample notification 1
+        String eventId1 = "8492iF6Eu41dwGCB1yUG";
+        String title1 = "Event Reminder";
+        String message1 = "Don't forget to join the upcoming event!";
+        notificationHelper.addNotificationToDevice(deviceId, eventId1, title1, message1);
+
+        // Sample notification 2
+        String eventId2 = "E4sUpzhqDYcazajGbdlW";
+        String title2 = "Special Offer";
+        String message2 = "Check out the exclusive offer available for event participants!";
+        notificationHelper.addNotificationToDevice(deviceId, eventId2, title2, message2);
+        */
+
         updateProfileButton.setOnClickListener(v -> {
             Intent profileIntent = new Intent(EntrantMain.this, ViewProfile.class);
             profileIntent.putExtra("DEVICE_ID", deviceId);
@@ -68,6 +100,31 @@ public class EntrantMain extends AppCompatActivity {
             profileIntent.putExtra("DEVICE_ID", deviceId);
             startActivity(profileIntent);
         });
+    }
+
+    public void createNotificationChannel(Context context) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationManager notificationManager = context.getSystemService(NotificationManager.class);
+            if (notificationManager != null && notificationManager.getNotificationChannel("default") == null) {
+                CharSequence name = "Default Channel";
+                String description = "Channel for default notifications";
+                int importance = NotificationManager.IMPORTANCE_DEFAULT;
+                NotificationChannel channel = new NotificationChannel("default", name, importance);
+                channel.setDescription(description);
+                notificationManager.createNotificationChannel(channel);
+            }
+        }
+    }
+
+    private void requestNotificationPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) { // API 33+
+            if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.POST_NOTIFICATIONS)
+                    != PackageManager.PERMISSION_GRANTED) {
+                // Request the permission if not already granted
+                ActivityCompat.requestPermissions(this,
+                        new String[]{Manifest.permission.POST_NOTIFICATIONS}, 1);
+            }
+        }
     }
 
     /**
